chore: validations schema update for settings
