--- packages/common/src/validations/address_book.ts
@@ -1,8 +1,0 @@
-import { z } from "zod";
-
-export const InsertAddressBook = z.object({
-    name: z.string(),
-    publicKey: z.string(),
-    chain: z.string(),
-    tag: z.string().optional(),
-});
