+++ packages/common/src/validations/index.ts
@@ -12,4 +12,4 @@ export * from "./notification";
 export * from "./buy";
 export * from "./locale";
 export * from "./settings";
-export * from "./address_book";
+export * from "./book";
